Movement behaviours implementation for IOOD and Boulder Beetles
This implements a base class called MovementHandler which
can be overridden to provide custom movement behaviour for
monsters. The ProjectileMovement class overrides this to
provide a base implementation for IOOD and Boulder Beetle.
OrbMovement and BoulderMovement implement the behaviours
respectively for those monsters.

+++ crawl-ref/source/player-act.cc
@@ -721,7 +721,7 @@ void player::attacking(actor *other, bool ranged)
     if (ranged || mons_is_firewood((monster*) other))
         return;
 
-    const int chance = pow(3, player_mutation_level(MUT_BERSERK) - 1);
+    const int chance = pow((double)3, player_mutation_level(MUT_BERSERK) - 1);
     if (player_mutation_level(MUT_BERSERK) && x_chance_in_y(chance, 100))
         go_berserk(false);
 }
